detect when tests delete important files (e.g. by cleaning the git)
Previous behaviour: If a test deletes .tmp-tasks, all the remaining tasks will be skipped and the build will succeed.

+++ prepare-and-run-ci-tests.sh
@@ -23,17 +23,19 @@ export ENVVAR_STORE="$(pwd)/.tmp-envvar_store"
 [ -e "$ENVVAR_STORE" ] && rm "$ENVVAR_STORE"
 ENVVAR_STORE_ALL=".tmp-envvar_store-all"
 [ -e "$ENVVAR_STORE_ALL" ] && rm "$ENVVAR_STORE_ALL"
+touch "$ENVVAR_STORE_ALL"
+TASK_LIST=".tmp-tasks"
 
 echo "Gathering task list..."
 
 find -regex ".*/ci-[0-9]+-[^/ ]+\.[^/ ]+" | awk -F "/" '{print $NF " " $0}' \
 	| sort \
 	| cut -d " " -f2- \
-	>.tmp-tasks
+	>"$TASK_LIST"
 if ! (($DONT_TEST)) ; then
-	find -name test-ci*.sh | sort >>.tmp-tasks
+	find -name test-ci*.sh | sort >>"$TASK_LIST"
 fi
-TASK_COUNT="$(wc -l .tmp-tasks | cut -d " " -f1)"
+TASK_COUNT="$(wc -l "$TASK_LIST" | cut -d " " -f1)"
 echo "Found $TASK_COUNT tasks."
 
 TASK_NUM=0
@@ -52,12 +54,17 @@ while read script ; do
 		exit $EXITCODE
 	)
 
+	if [ ! -e "$ENVVAR_STORE_ALL" -o ! -e "$TASK_LIST" ] ; then
+		echo "\n\nERROR: $ENVVAR_STORE_ALL and/ or $TASK_LIST have been deleted!" >&2
+		exit 1
+	fi
+
 	[ -e "$ENVVAR_STORE" ] \
 		&& cat "$ENVVAR_STORE" \
 		&& cat "$ENVVAR_STORE" >>"$ENVVAR_STORE_ALL" \
 		&& source "$ENVVAR_STORE" \
 		&& rm "$ENVVAR_STORE"
-done <.tmp-tasks
+done <"$TASK_LIST"
 
 echo
 echo
